Update OutfitterScripting.lua to use new spell tracking function
Previously the fish tracking enabled option was checking to see if fish tracking was already on and maintaining it, but was overwriting and not restoring any other kind of tracking. Changed the script to use the new GetCurrentSpellTrackingEnabled() function to identify what tracking was previously active (if any) and restore the state when the fishing outfit was removed.
I expect the original script was designed for Retail, where you can have multiple kinds of tracking together.